+++ src/RingBufferPlus/RingBufferExtension.cs
@@ -36,7 +36,7 @@ namespace RingBufferPlus
         /// <param name="loggerFactory">The logger factory to create a logger.</param>
         /// <returns><see cref="IRingBufferBuilder{T}"/>.</returns>
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "ByDesign")]
-        public static IRingBufferBuilder<T> New(string? buffername, ILoggerFactory loggerFactory)
+        public static IRingBufferBuilder<T> New(string buffername, ILoggerFactory loggerFactory)
         {
             if (buffername is null)
             {